+++ packages/adapter-interfaces/src/types/Password.ts
@@ -1,9 +1,16 @@
-export interface PasswordResponse {
-  valid: boolean;
-  message: string;
-}
+import { z } from "@hono/zod-openapi";
 
-export interface PasswordParams {
-  user_id: string;
-  password: string;
-}
+export const passwordInsertSchema = z.object({
+  user_id: z.string(),
+  password: z.string(),
+});
+
+export type PasswordInsert = z.infer<typeof passwordInsertSchema>;
+
+export const passwordSchema = z.object({
+  ...passwordInsertSchema.shape,
+  created_at: z.string(),
+  updated_at: z.string(),
+});
+
+export type Password = z.infer<typeof passwordSchema>;