+++ packages/kysely/src/passwords/index.ts
@@ -1,4 +1,4 @@
-import { validate } from "./validate";
+import { get } from "./get";
 import { create } from "./create";
 import { update } from "./update";
 import { Kysely } from "kysely";
@@ -9,6 +9,6 @@ export function createPasswordAdapter(db: Kysely<Database>): PasswordsAdapter {
   return {
     create: create(db),
     update: update(db),
-    validate: validate(db),
+    get: get(db),
   };
 }