+++ packages/kysely/src/passwords/update.ts
@@ -1,20 +1,17 @@
 import { Kysely } from "kysely";
-import bcrypt from "bcryptjs";
 import { Database } from "../db";
-import { PasswordParams } from "@authhero/adapter-interfaces";
+import { PasswordInsert } from "@authhero/adapter-interfaces";
 
 export function update(db: Kysely<Database>) {
-  return async (tenant_id: string, params: PasswordParams) => {
-    const passwordHash = bcrypt.hashSync(params.password, 10);
-
+  return async (tenant_id: string, password: PasswordInsert) => {
     const results = await db
       .updateTable("passwords")
       .set({
-        password: passwordHash,
+        password: password.password,
         updated_at: new Date().toISOString(),
       })
       .where("tenant_id", "=", tenant_id)
-      .where("user_id", "=", params.user_id)
+      .where("user_id", "=", password.user_id)
       .execute();
 
     return results.length === 1;